move linux_get_libc_version to conda.common and clean up imports
