+++ conda/_vendor/__init__.py
@@ -7,4 +7,3 @@ until conda 5.0 when conda will be isolated in its own private environment.
 Introduction of dependencies for the 4.x series is discussed in
 https://github.com/conda/conda/issues/2825.
 """
-from __future__ import absolute_import, division, print_function
